+++ src/components/Editor/editorClassnames.ts
@@ -1,4 +1,4 @@
-export const classnames = {
+export const editorClassnames = {
   editorContainer: 'ql-container',
   editor: 'ql-editor',
   toolbar: 'ql-toolbar',
